cleancode 魔鬼数字

+++ attention/fused_infer_attention_score/op_host/arch32/fused_infer_attention_score_tiling_check_feature.cpp
@@ -22,12 +22,12 @@ using std::pair;
 using namespace ge;
 using namespace AscendC;
 namespace optiling {
-constexpr size_t ALIGN_SIZE_16 = 16;
-constexpr size_t ALIGN_SIZE_32 = 32;
-constexpr size_t HEAD_NUM_ONE = 1;
-constexpr size_t HEAD_DIM_64 = 64;
-constexpr size_t HEAD_DIM_128 = 128;
-constexpr size_t HEAD_DIM_192 = 192;
+constexpr uint32_t ALIGN_SIZE_16 = 16;
+constexpr uint32_t ALIGN_SIZE_32 = 32;
+constexpr uint32_t HEAD_NUM_ONE = 1;
+constexpr uint32_t HEAD_DIM_64 = 64;
+constexpr uint32_t HEAD_DIM_128 = 128;
+constexpr uint32_t HEAD_DIM_192 = 192;
 
 ge::graphStatus FiaTilingCheck::CheckFeatureNoQuantDtype() const
 {
